+++ src/app/views/partial/cookies-disabled.v.php
@@ -1,0 +1,5 @@
+<div id="cookies_disabled" class="cookies-disabled panel panel-danger d-none">
+	<div class="panel-heading">
+		<strong><span class="glyphicon glyphicon-alert"></span> &nbsp; <?=$this->t('--cookies-disabled--') ?></strong>
+	</div>
+</div>